[MouseWithoutBorders] - split "Common.Log.cs" into "Logger.cs" - #34126
